+++ master/buildbot/test/integration/test_slave_comm.py
@@ -26,7 +26,7 @@ from buildbot.process import factory
 from buildbot.status import master
 from buildbot.test.fake import fakemaster
 from buildbot.test.util import compat
-from buildbot.test.util.flaky import flaky
+from buildbot.test.util.decorators import flaky
 from buildbot.util.eventual import eventually
 from twisted.cred import credentials
 from twisted.internet import defer
